requirements.txt
certifi==2020.6.20
chardet==3.0.4
configobj==5.0.6
idna==2.10
pycairo==1.16.2
pycrypto==2.6.1
pyTelegramBotAPI==3.7.3
python-xlib==0.20
pyxdg==0.25
requests==2.7.0
setproctitle==1.1.10
six==1.15.0
telebot
python-telegram-bot
urllib3==1.25.10
ipython
matplotlib
torch
scikit-image
numpy
python-dotenv
